requirements.txt
Flask
Flask-Mail
opencv-python
face_recognition
numpy
Pillow
